+++ Sources/Monstore/MemoryCache/TTLPriorityLRUQueue.swift
@@ -1,5 +1,5 @@
 //
-//  LRUQueueWithTTL.swift
+//  TTLPriorityLRUQueue.swift
 //  Monstore
 //
 //  Created by Larkin on 2025/5/8.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-class LRUQueueWithTTL<Key: Hashable, Value> {
+class TTLPriorityLRUQueue<Key: Hashable, Value> {
     let capacity: Int
     fileprivate var ttlQueue: Heap<Node>
     fileprivate var lruQueue: PriorityLRUQueue<Key, Node>
@@ -41,9 +41,9 @@ class LRUQueueWithTTL<Key: Hashable, Value> {
     }
 }
 
-extension LRUQueueWithTTL {
+extension TTLPriorityLRUQueue {
     @discardableResult
-    func unsafeSet(value: Value, for key: Key, expiredIn duration: TimeInterval) -> Value? {
+    func unsafeSet(value: Value, for key: Key, priority: Double = .zero, expiredIn duration: TimeInterval = .infinity) -> Value? {
         let now = CPUTimeStamp.now()
 
         // Check if the key already exists in the LRU queue
@@ -51,19 +51,19 @@ extension LRUQueueWithTTL {
             // If the existing entry has expired, remove it and insert the new value into the TTL heap
             if existing.expirationTimeStamp < now {
                 _ = unsafeRemoveValue(for: key)
-                return unsafeSetToTTLQueue(now: now, value: value, for: key, expiredIn: duration)
+                return unsafeSetToTTLQueue(now: now, value: value, for: key, priority: priority, expiredIn: duration)
             } else {
                 // If the existing entry is still valid, update its position in the LRU queue
-                return unsafeSetToLRUQueue(now: now, value: value, for: key, expiredIn: duration)
+                return unsafeSetToLRUQueue(now: now, value: value, for: key, priority: priority, expiredIn: duration)
             }
         } else {
             // If the key does not exist, check if the TTL heap root has expired
             if let ttlRoot = ttlQueue.root?.expirationTimeStamp, ttlRoot < now {
                 // If the root of TTL heap is expired, insert into the TTL heap
-                return unsafeSetToTTLQueue(now: now, value: value, for: key, expiredIn: duration)
+                return unsafeSetToTTLQueue(now: now, value: value, for: key, priority: priority, expiredIn: duration)
             }
             // Otherwise, insert into the LRU queue
-            return unsafeSetToLRUQueue(now: now, value: value, for: key, expiredIn: duration)
+            return unsafeSetToLRUQueue(now: now, value: value, for: key, priority: priority, expiredIn: duration)
         }
     }
     
@@ -88,22 +88,22 @@ extension LRUQueueWithTTL {
     }
 }
 
-private extension LRUQueueWithTTL {
-    func unsafeSetToTTLQueue(now: CPUTimeStamp, value: Value, for key: Key, expiredIn duration: TimeInterval) -> Value? {
+private extension TTLPriorityLRUQueue {
+    func unsafeSetToTTLQueue(now: CPUTimeStamp, value: Value, for key: Key, priority: Double, expiredIn duration: TimeInterval) -> Value? {
         var res: Value? = nil
         let node = Node(key: key, value: value, expirationTimeStamp: now + duration)
         if let pop = ttlQueue.insert(node, force: true) {
             _=lruQueue._removeValue(for: pop.key)
             res = pop.value
         }
-        _=lruQueue._setValue(node, for: key, with: 0)
+        _=lruQueue._setValue(node, for: key, with: priority)
         return res
     }
     
-    func unsafeSetToLRUQueue(now: CPUTimeStamp, value: Value, for key: Key, expiredIn duration: TimeInterval) -> Value? {
+    func unsafeSetToLRUQueue(now: CPUTimeStamp, value: Value, for key: Key, priority: Double, expiredIn duration: TimeInterval) -> Value? {
         var res: Value? = nil
         let node = Node(key: key, value: value, expirationTimeStamp: now + duration)
-        if let pop = lruQueue._setValue(node, for: key, with: 0), let ttlIndex = pop.ttlIndex {
+        if let pop = lruQueue._setValue(node, for: key, with: priority), let ttlIndex = pop.ttlIndex {
             _=ttlQueue.remove(at: ttlIndex)
             res = pop.value
         }
@@ -112,7 +112,7 @@ private extension LRUQueueWithTTL {
     }
 }
 
-private extension LRUQueueWithTTL {
+private extension TTLPriorityLRUQueue {
     class Node {
         let key: Key
         let value: Value